refactor: refactor icon imports to centralized index file for cleaner code
