+++ apps/web/src/shared/icons/index.ts
@@ -1,0 +1,25 @@
+export { default as AgreeIcon } from './AgreeIcon.svg?react';
+export { default as ArrowBackIcon } from './ArrowBackIcon.svg?react';
+export { default as ArrowDownIcon } from './ArrowDownIcon.svg?react';
+export { default as ArrowUpIcon } from './ArrowUpIcon.svg?react';
+export { default as CategoryOffIcon } from './CategoryOffIcon.svg?react';
+export { default as CategoryOnIcon } from './CategoryOnIcon.svg?react';
+export { default as DisagreeIcon } from './DisagreeIcon.svg?react';
+export { default as HeartBlackOffIcon } from './HeartBlackOffIcon.svg?react';
+export { default as HeartOnIcon } from './HeartOnIcon.svg?react';
+export { default as HeartWhiteOffIcon } from './HeartWhiteOffIcon.svg?react';
+export { default as HomeOffIcon } from './HomeOffIcon.svg?react';
+export { default as HomeOnIcon } from './HomeOnIcon.svg?react';
+export { default as MainArrowIcon } from './MainArrowIcon.svg?react';
+export { default as NoRankChangeIcon } from './NoRankChangeIcon.svg?react';
+export { default as ProfileOffIcon } from './ProfileOffIcon.svg?react';
+export { default as ProfileOnIcon } from './ProfileOnIcon.svg?react';
+export { default as RankingOffIcon } from './RankingOffIcon.svg?react';
+export { default as RankingOnIcon } from './RankingOnIcon.svg?react';
+export { default as ReviewOffIcon } from './ReviewOffIcon.svg?react';
+export { default as ReviewOnIcon } from './ReviewOnIcon.svg?react';
+export { default as SearchIcon } from './SearchIcon.svg?react';
+export { default as SomesayIcon } from './SomesayIcon.svg?react';
+export { default as Star16Icon } from './Star16Icon.svg?react';
+export { default as ToggleDownIcon } from './ToggleDownIcon.svg?react';
+export { default as YoutubeIcon } from './YoutubeIcon.svg?react';
